+++ sdk/appconfiguration/azure-appconfiguration-provider/azure/appconfiguration/provider/aio/_azureappconfigurationproviderasync.py
@@ -5,13 +5,12 @@
 # -------------------------------------------------------------------------
 import json
 import datetime
-from threading import Lock
 import logging
 from typing import (
     Any,
     Awaitable,
-    Callable,
     Dict,
+    Callable,
     Mapping,
     Optional,
     overload,
@@ -19,11 +18,6 @@ from typing import (
     Tuple,
     TYPE_CHECKING,
     Union,
-    Iterator,
-    KeysView,
-    ItemsView,
-    ValuesView,
-    TypeVar,
 )
 from azure.appconfiguration import (  # type:ignore # pylint:disable=no-name-in-module
     FeatureFlagConfigurationSetting,
@@ -37,23 +31,20 @@ from .._models import AzureAppConfigurationKeyVaultOptions, SettingSelector
 from .._constants import (
     FEATURE_MANAGEMENT_KEY,
     FEATURE_FLAG_KEY,
-    EMPTY_LABEL,
+)
+from .._azureappconfigurationproviderbase import (
+    AzureAppConfigurationProviderBase,
+    get_headers,
+    delay_failure,
+    is_json_content_type,
 )
 from ._async_client_manager import AsyncConfigurationClientManager
-from .._azureappconfigurationprovider import (
-    _get_headers,
-    _RefreshTimer,
-    _build_sentinel,
-    _delay_failure,
-    _is_json_content_type,
-)
 from .._user_agent import USER_AGENT
 
 if TYPE_CHECKING:
     from azure.core.credentials_async import AsyncTokenCredential
 
 JSON = Mapping[str, Any]
-_T = TypeVar("_T")
 
 
 @overload
@@ -223,7 +214,7 @@ async def load(*args, **kwargs) -> "AzureAppConfigurationProvider":
     provider = await _buildprovider(connection_string, endpoint, credential, uses_key_vault=uses_key_vault, **kwargs)
     # Discovering replicas outside of init as it's async
     await provider._replica_client_manager.setup_initial_clients()  # pylint:disable=protected-access
-    headers = _get_headers(
+    headers = get_headers(
         kwargs.pop("headers", {}),
         "Startup",
         provider._replica_client_manager.get_client_count() - 1,  # pylint:disable=protected-access
@@ -235,7 +226,7 @@ async def load(*args, **kwargs) -> "AzureAppConfigurationProvider":
     try:
         await provider._load_all(headers=headers)  # pylint:disable=protected-access
     except Exception as e:
-        _delay_failure(start_time)
+        delay_failure(start_time)
         raise e
     return provider
 
@@ -304,7 +295,7 @@ async def _resolve_keyvault_reference(
     raise ValueError("No Secret Client found for Key Vault reference %s" % (vault_url))
 
 
-class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: disable=too-many-instance-attributes
+class AzureAppConfigurationProvider(AzureAppConfigurationProviderBase):  # pylint: disable=too-many-instance-attributes
     """
     Provides a dictionary-like interface to Azure App Configuration settings. Enables loading of sets of configuration
     settings from Azure App Configuration into a Python application. Enables trimming of prefixes from configuration
@@ -312,8 +303,7 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
     """
 
     def __init__(self, **kwargs: Any) -> None:
-        endpoint = kwargs.pop("endpoint", None)
-        self._origin_endpoint = endpoint
+        super().__init__(**kwargs)
 
         if "user_agent" in kwargs:
             user_agent = kwargs.pop("user_agent") + " " + USER_AGENT
@@ -329,7 +319,7 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
 
         self._replica_client_manager = AsyncConfigurationClientManager(
             connection_string=kwargs.pop("connection_string", None),
-            endpoint=endpoint,
+            endpoint=kwargs.pop("endpoint", None),
             credential=kwargs.pop("credential", None),
             user_agent=user_agent,
             retry_total=kwargs.pop("retry_total", 2),
@@ -339,38 +329,11 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
             max_backoff_sec=max_backoff,
             **kwargs,
         )
-        self._dict: Dict[str, Any] = {}
         self._secret_clients: Dict[str, SecretClient] = {}
-        self._selects: List[SettingSelector] = kwargs.pop(
-            "selects", [SettingSelector(key_filter="*", label_filter=EMPTY_LABEL)]
-        )
-
-        trim_prefixes: List[str] = kwargs.pop("trim_prefixes", [])
-        self._trim_prefixes: List[str] = sorted(trim_prefixes, key=len, reverse=True)
-
-        refresh_on: List[Tuple[str, str]] = kwargs.pop("refresh_on", None) or []
-        self._refresh_on: Mapping[Tuple[str, str], Optional[str]] = {_build_sentinel(s): None for s in refresh_on}
-        self._refresh_timer: _RefreshTimer = _RefreshTimer(**kwargs)
         self._on_refresh_success: Optional[Callable] = kwargs.pop("on_refresh_success", None)
         self._on_refresh_error: Optional[Union[Callable[[Exception], Awaitable[None]], None]] = kwargs.pop(
             "on_refresh_error", None
         )
-        self._keyvault_credential = kwargs.pop("keyvault_credential", None)
-        self._secret_resolver = kwargs.pop("secret_resolver", None)
-        self._keyvault_client_configs = kwargs.pop("keyvault_client_configs", {})
-        self._uses_key_vault = (
-            self._keyvault_credential is not None
-            or (self._keyvault_client_configs is not None and len(self._keyvault_client_configs) > 0)
-            or self._secret_resolver is not None
-        )
-        self._feature_flag_enabled = kwargs.pop("feature_flag_enabled", False)
-        self._feature_flag_selectors = kwargs.pop("feature_flag_selectors", [SettingSelector(key_filter="*")])
-        self._refresh_on_feature_flags: Mapping[Tuple[str, str], Optional[str]] = {}
-        self._feature_flag_refresh_timer: _RefreshTimer = _RefreshTimer(**kwargs)
-        self._feature_flag_refresh_enabled = kwargs.pop("feature_flag_refresh_enabled", False)
-        self._feature_filter_usage: Mapping[str, bool] = {}
-        self._update_lock = Lock()
-        self._refresh_lock = Lock()
 
     async def refresh(self, **kwargs) -> None:
         if not self._refresh_on and not self._feature_flag_refresh_enabled:
@@ -392,7 +355,7 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
             await self._replica_client_manager.refresh_clients()
             active_clients = self._replica_client_manager.get_active_clients()
 
-            headers = _get_headers(
+            headers = get_headers(
                 kwargs.pop("headers", {}),
                 "Watch",
                 self._replica_client_manager.get_client_count() - 1,
@@ -502,19 +465,10 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
             "Failed to load configuration settings. No Azure App Configuration stores successfully loaded from."
         )
 
-    def _process_key_name(self, config):
-        trimmed_key = config.key
-        # Trim the key if it starts with one of the prefixes provided
-        for trim in self._trim_prefixes:
-            if config.key.startswith(trim):
-                trimmed_key = config.key[len(trim) :]
-                break
-        return trimmed_key
-
     async def _process_key_value(self, config):
         if isinstance(config, SecretReferenceConfigurationSetting):
             return await _resolve_keyvault_reference(config, self)
-        if _is_json_content_type(config.content_type) and not isinstance(config, FeatureFlagConfigurationSetting):
+        if is_json_content_type(config.content_type) and not isinstance(config, FeatureFlagConfigurationSetting):
             # Feature flags are of type json, but don't treat them as such
             try:
                 return json.loads(config.value)
@@ -522,80 +476,6 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
                 # If the value is not a valid JSON, treat it like regular string value
                 return config.value
         return config.value
-
-    def __getitem__(self, key: str) -> Any:
-        # pylint:disable=docstring-missing-param,docstring-missing-return,docstring-missing-rtype
-        """
-        Returns the value of the specified key.
-        """
-        with self._update_lock:
-            return self._dict[key]
-
-    def __iter__(self) -> Iterator[str]:
-        return self._dict.__iter__()
-
-    def __len__(self) -> int:
-        return len(self._dict)
-
-    def __contains__(self, __x: object) -> bool:
-        # pylint:disable=docstring-missing-param,docstring-missing-return,docstring-missing-rtype
-        """
-        Returns True if the configuration settings contains the specified key.
-        """
-        return self._dict.__contains__(__x)
-
-    def keys(self) -> KeysView[str]:
-        """
-        Returns a list of keys loaded from Azure App Configuration.
-
-        :return: A list of keys loaded from Azure App Configuration.
-        :rtype: KeysView[str]
-        """
-        with self._update_lock:
-            return self._dict.keys()
-
-    def items(self) -> ItemsView[str, Union[str, Mapping[str, Any]]]:
-        """
-        Returns a set-like object of key-value pairs loaded from Azure App Configuration. Any values that are Key Vault
-         references will be resolved.
-
-        :return: A set-like object of key-value pairs loaded from Azure App Configuration.
-        :rtype: ItemsView[str, Union[str, Mapping[str, Any]]]
-        """
-        with self._update_lock:
-            return self._dict.items()
-
-    def values(self) -> ValuesView[Union[str, Mapping[str, Any]]]:
-        """
-        Returns a list of values loaded from Azure App Configuration. Any values that are Key Vault references will be
-        resolved.
-
-        :return: A list of values loaded from Azure App Configuration. The values are either Strings or JSON objects,
-         based on there content type.
-        :rtype: ValuesView[Union[str, Mapping[str, Any]]]
-        """
-        with self._update_lock:
-            return (self._dict).values()
-
-    @overload
-    def get(self, key: str, default: None = None) -> Union[str, JSON, None]: ...
-
-    @overload
-    def get(self, key: str, default: Union[str, JSON, _T]) -> Union[str, JSON, _T]:  # pylint: disable=signature-differs
-        ...
-
-    def get(self, key: str, default: Optional[Union[str, JSON, _T]] = None) -> Union[str, JSON, _T, None]:
-        """
-        Returns the value of the specified key. If the key does not exist, returns the default value.
-
-        :param str key: The key of the value to get.
-        :param default: The default value to return.
-        :type: str or None
-        :return: The value of the specified key.
-        :rtype: Union[str, JSON]
-        """
-        with self._update_lock:
-            return self._dict.get(key, default)
 
     def __eq__(self, other: Any) -> bool:
         if not isinstance(other, AzureAppConfigurationProvider):
@@ -605,9 +485,6 @@ class AzureAppConfigurationProvider(Mapping[str, Union[str, JSON]]):  # pylint: 
         if self._trim_prefixes != other._trim_prefixes:
             return False
         return self._replica_client_manager == other._replica_client_manager
-
-    def __ne__(self, other: Any) -> bool:
-        return not self == other
 
     async def close(self) -> None:
         """
